+++ R/RWeka_CBA.R
@@ -1,0 +1,102 @@
+# interface to rWeka
+
+# M5Rules is only for binary classifier
+# OneR does not make too much sense with pre-discretized data
+
+#reassamble a data.frame from transactions
+# This combines variables and renames variables and values to numbers
+.trans2DF <- function(trans) {
+  ii <- itemInfo(trans)
+  if(is.null(ii$variables)) stop("Transaction itemInfo does not contain variable information.")
+
+  vars <- unique(ii$variables)
+
+  df <- lapply(vars, FUN = function(v) {
+      t_v <- trans[,ii$variables == v]
+      factor(apply(as(t_v, "ngCMatrix"), MARGIN = 2, which), labels = (colnames(t_v)))
+    })
+
+  df <- as.data.frame(df)
+  colnames(df) <- vars
+  df
+}
+
+
+
+.rules_RWeka <- function(formula, data, what = RWeka::JRip, control = NULL,
+  disc.method = "mdlp") {
+
+  if(is.null(control)) control <- RWeka::Weka_control()
+  formula <- as.formula(formula)
+
+  # prepare data
+  disc_info <- NULL
+  if(is(data, "data.frame")){
+    data <- discretizeDF.supervised(formula, data, method = disc.method)
+    disc_info <- lapply(data, attr, "discretized:breaks")
+  }
+
+  ### FIXME: this is probably not necessary...
+
+  # convert to transactions for rule mining
+  trans <- as(data, "transactions")
+
+  parsedFormula <- .parseformula(formula, trans)
+
+  # convert it back
+  data <- .trans2DF(trans)
+
+  classifier <- what(formula, data = data, control = control)
+
+  # extract rules
+  rules <- rJava::.jcall(classifier$classifier, "S", "toString")
+  if(substr(rules[1], 1, 4) == "JRIP") { ### RIPPER
+    rule_sep <- '\\s+=>\\s+'
+  }else{ ### PART
+    rules <- gsub(' AND\\n', ' ', x = rules, ) # replace AND used in PART
+    rule_sep <- ':\\s+'
+  }
+
+  rules <- strsplit(rules, '\\n+')[[1]]
+  rules <- rules[-c(1,2, length(rules))]
+  rules <- strsplit(rules, rule_sep)
+
+  lhs <- sapply(rules, '[', 1)
+  rhs <- sapply(rules, '[', 2)
+
+  ilabels <- itemLabels(trans)
+  n <- length(rules)
+  m <- nitems(trans)
+
+  mat <- matrix(FALSE, nrow = n, ncol = m, dimnames = list(rows = NULL, cols = ilabels))
+  for(i in 1:length(ilabels)) mat[grep(ilabels[i], lhs, fixed = TRUE), i] <- TRUE
+  lhs <- as(mat, "itemMatrix")
+
+  mat <- matrix(FALSE, nrow = n, ncol = m, dimnames = list(rows = NULL, cols = ilabels))
+  for(i in 1:length(ilabels)) mat[grep(ilabels[i], rhs, fixed = TRUE), i] <- TRUE
+  rhs <- as(mat, "itemMatrix")
+
+  rules <- new("rules", lhs = lhs, rhs = rhs)
+  quality(rules) <- interestMeasure(rules, measure = c("support", "confidence"),
+    transactions = trans)
+
+  # assemble classifier
+  structure(list(
+    rules = rules,
+    class = parsedFormula$class_name,
+    default = LIST(rhs(tail(rules, 1)))[[1]], ### FIXME: last rule is the default rules
+    #default = names(max(table(data))), ### FIXME: majority class
+    discretization = disc_info,
+    formula = formula,
+    method = "first",
+    description = paste("RWeka classifier", attr(what, "meta")$name)
+  ),
+    class = "CBA"
+  )
+}
+
+RIPPER <- function(formula, data, control = NULL, disc.method = "mdlp")
+  .rules_RWeka(formula, data, RWeka::JRip, control, disc.method)
+
+PART_CBA <- function(formula, data, control = NULL, disc.method = "mdlp")
+  .rules_RWeka(formula, data, RWeka::PART, control, disc.method)