+++ NEWS.md
@@ -1,6 +1,7 @@
 # arulesCBA 1.1.6.1 (2020-xx-xx)
 * added CMAR.
 * added FOIL.
+* added RIPPER and PART (via RWeka).
 
 # arulesCBA 1.1.6 (2020-01-05)